Fix a -Werror=conversion break I pushed, and how it got past me
`sum / count` with a size_t divisor. My standalone g++ build passes
-Wconversion but not -Werror, and I grepped its output for "error" alone — so
a warning that CMake turns into a hard failure read as clean. ctest then passed
because the target had failed to rebuild and it ran the previous binary.

Both halves are worth stating: a build check that greps only for errors is not
a build check when the project builds with -Werror, and a green ctest after a
failed compile is a green test run of the last code that compiled.

+++ motionwave/core/test/granular_delay_tests.cpp
@@ -491,7 +491,7 @@ MW_TEST("V4: the loop is bounded at every feedback, topology and cross setting")
           const double v = static_cast<double>(tail.left[i]);
           sum += v * v;
         }
-        return std::sqrt(sum / count);
+        return std::sqrt(sum / static_cast<double>(count));
       };
       const std::size_t second = static_cast<std::size_t>(kRate);
       const double early = rmsOver(0, second);
